% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calheatmapR.R
\name{calheatmapR}
\alias{calheatmapR}
\title{calheatmapR for plotting a calendar heatmap}
\source{
D3.js was created by Mike Bostock, see \url{http://d3js.org}, and
cal-heatmap.js was created by Wan Qi Chen, see \url{https://github.com/kamisama/cal-heatmap}
}
\usage{
calheatmapR(data, width = NULL, height = NULL)
}
\arguments{
\item{data}{a list object, each element's name is a timestamp in seconds, with
a corresponding number, see \href{http://kamisama.github.io/cal-heatmap/#data-format}{here}
for details.}

\item{width}{width of element}

\item{height}{height of element}
}
\description{
R interface to interactive plotting using the
\href{https://github.com/kamisama/cal-heatmap}{cal-heatmap} Javascript library
}

