% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/domains.R
\name{chDomain}
\alias{chDomain}
\title{Calendar Heatmap domain}
\usage{
chDomain(calheatmapR, domain = c("hour", "day", "week", "month", "year"),
  subDomain = c("min", "x_min", "hour", "x_hour", "day", "x_day", "week",
  "x_week", "month", "x_month"), start = NULL, range = 12, cellSize = 10,
  cellPadding = 2, cellRadius = 0, gutter = 2, margin = c(0, 0, 0, 0),
  colLimit = NULL, rowLimit = NULL, dynamicDimension = TRUE,
  verticalOrientation = FALSE)
}
\arguments{
\item{calheatmapR}{Calendar Heatmap to configure options for}

\item{domain}{Type of domain}

\item{subDomain}{Type of subDomain (must always be smaller than domain)}

\item{start}{Date, as a string, in YYYY-MM-DD format}

\item{range}{Number of domains to display}

\item{cellSize}{Size of each domain (in pixels)}

\item{cellPadding}{Space between each subDomain cell (in pixels)}

\item{cellRadius}{Border radius of subDomain cells, for rounder corners (in pixels)}

\item{gutter}{Space between each domain}

\item{margin}{Margins around each domain, either integer (0) or vector
c(0,0) or c(0,0,0,0), all of which produce the same. c(0,1) would create margins
of c(0,1,0,1) (in pixels)}

\item{colLimit}{Control the number of columns to split the domains into}

\item{rowLimit}{Control the number of rows to split the domains into}

\item{dynamicDimension}{Enable/disable dynamic width and height of domains}

\item{verticalOrientation}{Display Calendar Heatmap vertically}
}
\value{
a calheatmapR with customised domain options
}
\description{
Configure options with respect to the domains that make up a Calendar Heatmap
}

