% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/labels.R
\name{chLabel}
\alias{chLabel}
\title{Calendar Heatmap label}
\usage{
chLabel(calheatmapR, position = c("bottom", "left", "top", "right"),
  align = c("center", "right", "left"), rotate = NULL, width = 100,
  offset = list(x = 0, y = 0), height = NULL, itemName = c("item",
  "items"))
}
\arguments{
\item{calheatmapR}{Calendar Heatmap to configure options for}

\item{position}{Position of the label, with respect to the domain}

\item{align}{Alignment of the label, with respect to the domain}

\item{rotate}{Rotation for a vertical label, default is NULL, but accepts "left"
or "right"}

\item{width}{Applied when label is rotated, defines width of label}

\item{offset}{Additional control for label positioning}

\item{height}{Height of the domain label (in pixels)}

\item{itemName}{Name of entity on calender being presented}
}
\value{
a calheatmapR with customised label options
}
\description{
Configure options with respect to the labels in a Calendar Heatmap
}

