% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/legend.R
\name{chLegend}
\alias{chLegend}
\title{Calendar Heatmap legend}
\usage{
chLegend(calheatmapR, legend = c(10, 20, 30, 40), display = TRUE,
  cellSize = 10, cellPadding = 2, margin = c(10, 0, 0, 0),
  verticalPosition = c("bottom", "top", "center"),
  horizontalPosition = c("left", "center", "right"),
  orientation = c("horizontal", "vertical"), colours = NULL)
}
\arguments{
\item{calheatmapR}{Calendar Heatmap to configure options for}

\item{legend}{Numeric vector, assigns each range of values to a colour}

\item{display}{Whether to display the legend}

\item{cellSize}{Size of the legend cells (in pixels)}

\item{cellPadding}{Padding between each legend cell (in pixels)}

\item{margin}{Margins around each legend cell (in pixels)}

\item{verticalPosition}{Vertical position of the legend}

\item{horizontalPosition}{Horizontal position of the legend}

\item{orientation}{Orientation of the legend}

\item{colours}{Set of colours to compute the heatmap colours, either array or
list.  List object would contain a 'min' and 'max' elements, with optional
elements of 'empty', 'base' and 'overflow'}
}
\value{
a calheatmapR with customised label options
}
\description{
Configure options with respect to the legend in a Calendar Heatmap
}
\note{
see \href{https://kamisama.github.io/cal-heatmap/#legend}{here} for options
}

