% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calheatmapR.R
\docType{data}
\name{pletcher}
\alias{pletcher}
\title{Todd Pletcher racehorse trainer performance}
\description{
List containing data about racehorse trainer Todd Pletcher, tracking the
percentage of horses his horses beat in races they entered, over a 3month
period
}
\keyword{dataset}

